+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.3-unofficial-mushymato.1](UNRELEASED)
+
+### Fixed
+- `ReflectionPropertyDescriptor` should use `MethodInfoExtensions.SafeCreateDelegate` to correctly handle property getters on value types.
+
 ## [0.6.3-unofficial-mushymato.0](https://github.com/Mushymato/StardewUI/releases/tag/0.6.3-unofficial-mushymato.0) - WIP
 
 ### Added